api-tests/requirements.txt
requests==2.31.0
pytest==7.4.3
pytest-html==4.1.1
pytest-json-report==1.5.0
python-dotenv==1.0.0
allure-pytest==2.13.2
jsonschema==4.20.0
faker==20.1.0
pyyaml==6.0.1